SearchOption의 SearchCount와 Offset 형식 int로 변경

+++ PlurCrawler/Search/Base/ISearchOption.cs
@@ -15,12 +15,12 @@ namespace PlurCrawler.Search.Base
         /// <summary>
         /// 검색할 갯수입니다.
         /// </summary>
-        ulong SearchCount { get; set; }
+        int SearchCount { get; set; }
 
         /// <summary>
         /// 페이지의 오프셋을 결정합니다. 예를 들어 4를 입력했다면 5번째 결과부터 출력됩니다.
         /// </summary>
-        ulong Offset { get; set; }
+        int Offset { get; set; }
 
         /// <summary>
         /// 검색할 언어 코드를 입력합니다.
