+++ PlurCrawler/Search/Services/Youtube/YoutubeSearchOption.cs
@@ -19,12 +19,12 @@ namespace PlurCrawler.Search.Services.Youtube
         /// <summary>
         /// 검색할 갯수입니다.
         /// </summary>
-        public ulong SearchCount { get; set; }
+        public int SearchCount { get; set; }
 
         /// <summary>
         /// 페이지의 오프셋을 결정합니다. 예를 들어 4를 입력했다면 5번째 결과부터 출력됩니다.
         /// </summary>
-        public ulong Offset { get; set; }
+        public int Offset { get; set; }
 
         /// <summary>
         /// 검색할 언어 코드를 입력합니다.
